M value tweak

+++ create_data_submission.py
@@ -393,7 +393,7 @@ def deuniqueWorker(input, output):
         # calculated when working with a modelling project where I was subsampling to 1000 sequences. In this
         # scenario the M was set to 4.
         # We should also take care that M doesn't go below 4, so we should use a max choice for the M
-        M_value = max(4, (0.004 * (len(deuniquedFasta)/2)))
+        M_value = max(4, int(0.004 * (len(deuniquedFasta)/2)))
         completedProcess = subprocess.run(
             [r'decompose', '-M', M_value, '--skip-gexf-files', '--skip-gen-figures', '--skip-gen-html', '--skip-check-input', '-o',
              MEDOutDir, pathToFile], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
